Fix slow tests with mock forex fixture
Since forex data is pulled with every request of `ensure_folio_exists` , all the basic functionality tests are drastically slowed down. This introduced a automatic pytest fixture to mock some of the expensive API calls that is utilized by default.

This can be explicitly turned off when trying to verify the real forex API calls.

Refs #28