Seed mock with constant date for caching
The `conftest.py` fixture caches the settlement schedule for a certain period. To ensure that we are within this schedule, the mock data needs to have a constant seed date to have expectable behavior with preset expected dates.

The `settlement_calculator` is updated to mark non-computed Action types as not calculated or '0' , in preparation for import statements to verify settlement dates.

The `settle-info` commands response is also updated accordingly, as it will only account for BUSINESS_DAY_SETTLE_ACTIONS now.

refs #31